new config and change driver execute modbus data

+++ driver/devicetype.go
@@ -27,7 +27,7 @@ type ProtocolConfig struct {
 
 type ConfigData struct {
 	// TODO: add your protocol config data
-	SlaveID int `json:"slaveID,omitempty"`
+	SlaveID int `json:"slaveID,omitempty"` // 从站ID
 	CommunicationMode `json:"communication"` // 通信模式:TCP/RTU 
 }
 type CommunicationMode struct{
@@ -46,9 +46,9 @@ type VisitorConfigData struct {
 	// TODO: add your visitor config data
 	DataType string `json:"dataType"` // 数据类型:enum:Int/Float/String
 	Register string `json:"register"` // 寄存器类型:enum:CoilRegister/HoldingRegister
-	Offset int `json:"offset"`        // 寄存器偏移量
+	Offset uint16 `json:"offset"`        // 寄存器偏移量
 	Scale float64 `json:"scale"`      // 数据缩放比例
 	IsSwap bool `json:"isSwap"`       // 是否交换字节
-	Limit int `json:"limit"`          // 读取数量
+	Limit uint16 `json:"limit"`          // 读取数量
 	IsRegisterSwap bool `json:"isRegisterSwap"` // 是否交换寄存器
 }
